refactor(st-09074): modularize delete executor
