+++ packages/tools/tests/data/relational/tools/delete-executor/result-shaping.suite.ts
@@ -1,0 +1,107 @@
+import { describe, expect, it, vi } from 'vitest';
+import type { TransactionContext } from '../../../../../src/data/relational/query/transaction.js';
+import { createMockManager, executeDelete } from './shared.js';
+
+describe('relational-delete > executor > result shaping', () => {
+  it('should execute a single DELETE and return rowCount', async () => {
+    const manager = createMockManager([{ affectedRows: 5 }]);
+
+    const result = await executeDelete(manager, {
+      table: 'users',
+      where: [{ column: 'status', operator: 'eq', value: 'expired' }],
+      vendor: 'postgresql',
+      connectionString: 'postgresql://localhost/test',
+    });
+
+    expect(result.rowCount).toBe(5);
+    expect(result.executionTime).toBeGreaterThanOrEqual(0);
+    expect(manager.execute).toHaveBeenCalledOnce();
+  });
+
+  it('should handle object-style result with rowCount property', async () => {
+    const manager = createMockManager({ rowCount: 3 });
+
+    const result = await executeDelete(manager, {
+      table: 'sessions',
+      where: [{ column: 'expired', operator: 'eq', value: true }],
+      vendor: 'postgresql',
+      connectionString: 'postgresql://localhost/test',
+    });
+
+    expect(result.rowCount).toBe(3);
+  });
+
+  it('should handle SQLite changes result', async () => {
+    const manager = createMockManager([{ changes: 2 }]);
+
+    const result = await executeDelete(manager, {
+      table: 'logs',
+      where: [{ column: 'id', operator: 'in', value: [1, 2] }],
+      vendor: 'sqlite',
+      connectionString: ':memory:',
+    });
+
+    expect(result.rowCount).toBe(2);
+  });
+
+  it('should handle null result', async () => {
+    const manager = createMockManager(null);
+
+    const result = await executeDelete(manager, {
+      table: 'users',
+      where: [{ column: 'id', operator: 'eq', value: 999 }],
+      vendor: 'postgresql',
+      connectionString: 'postgresql://localhost/test',
+    });
+
+    expect(result.rowCount).toBe(0);
+  });
+
+  it('should report softDeleted flag for soft-delete queries', async () => {
+    const manager = createMockManager([{ affectedRows: 1 }]);
+
+    const result = await executeDelete(manager, {
+      table: 'users',
+      where: [{ column: 'id', operator: 'eq', value: 1 }],
+      softDelete: { column: 'deleted_at' },
+      vendor: 'postgresql',
+      connectionString: 'postgresql://localhost/test',
+    });
+
+    expect(result.softDeleted).toBe(true);
+    expect(result.rowCount).toBe(1);
+  });
+
+  it('should report softDeleted as false for hard-delete', async () => {
+    const manager = createMockManager([{ affectedRows: 1 }]);
+
+    const result = await executeDelete(manager, {
+      table: 'users',
+      where: [{ column: 'id', operator: 'eq', value: 1 }],
+      vendor: 'postgresql',
+      connectionString: 'postgresql://localhost/test',
+    });
+
+    expect(result.softDeleted).toBe(false);
+  });
+
+  it('should use transaction context when provided', async () => {
+    const txExecute = vi.fn().mockResolvedValue([{ affectedRows: 1 }]);
+    const manager = createMockManager();
+    const tx = { execute: txExecute } as unknown as TransactionContext;
+
+    await executeDelete(
+      manager,
+      {
+        table: 'users',
+        where: [{ column: 'id', operator: 'eq', value: 1 }],
+        vendor: 'postgresql',
+        connectionString: 'postgresql://localhost/test',
+      },
+      { transaction: tx },
+    );
+
+    expect(txExecute).toHaveBeenCalledOnce();
+    expect(manager.execute).not.toHaveBeenCalled();
+  });
+});
