+++ planning/features/09-solid-micro-refactors-feature-plan.md
@@ -2,8 +2,8 @@
 
 **Epic Range:** EP-09 through EP-09
 **Status:** In Progress
-**Last Updated:** 2026-06-27
-**Active Story:** ST-09074 - Modularize Relational Delete Executor and Tests (Ready)
+**Last Updated:** 2026-06-28
+**Active Story:** ST-09074 - Modularize Relational Delete Executor and Tests (In Progress)
 
 ---
